Manager: split out public API into sub-module
